Updates and simple implementation
Continued development of the full P2P system with users which is located in the (user.py) file. However this system is still unusable and does not allow for messages across the users. Instead the user_simple.py was added which showcases a simple form of messaging between two hardcoded users